try reorganizing test target common dependencies
