+++ bench/CMakeLists.txt
@@ -17,14 +17,13 @@ add_dependencies(benchmarks utrecht_tiles)
 macro (add_valhalla_benchmark target_file)
   set(target_name benchmark-${target_file})
   add_executable(${target_name}
-    ${target_file}.cc
-    ${VALHALLA_SOURCE_DIR}/third_party/microtar/src/microtar.c)
+    ${target_file}.cc)
   set_target_properties(${target_name} PROPERTIES FOLDER "Benchmarks")
   set_target_properties(${target_name}
     PROPERTIES
     COMPILE_DEFINITIONS
     VALHALLA_SOURCE_DIR="${VALHALLA_SOURCE_DIR}/")
-  target_link_libraries(${target_name} valhalla benchmark::benchmark)
+  target_link_libraries(${target_name} valhalla_test benchmark::benchmark)
   add_dependencies(benchmarks ${target_name})
   add_dependencies(${target_name} utrecht_tiles)
   # Add a custom target running the benchmark
