+++ src/sif/bicyclecost.cc
@@ -10,7 +10,7 @@
 #include <cassert>
 
 #ifdef INLINE_TEST
-#include "test/test.h"
+#include "test.h"
 #include "worker.h"
 #include <random>
 #endif
