+++ src/sif/nocost.cc
@@ -8,7 +8,7 @@
 #include "sif/dynamiccost.h"
 
 #ifdef INLINE_TEST
-#include "test/test.h"
+#include "test.h"
 #include "worker.h"
 #include <random>
 #endif
