+++ src/sif/truckcost.cc
@@ -9,7 +9,7 @@
 #include <cassert>
 
 #ifdef INLINE_TEST
-#include "test/test.h"
+#include "test.h"
 #include "worker.h"
 #include <random>
 #endif
