+++ test/gurka/test_route.cc
@@ -1,10 +1,7 @@
 #include "gurka.h"
-#include "test/util/traffic_utils.h"
-
-#include <gtest/gtest.h>
+#include "test.h"
 
 using namespace valhalla;
-namespace vu = valhalla_tests::utils;
 
 /*************************************************************/
 TEST(Standalone, TruckRegression) {
@@ -265,17 +262,22 @@ protected:
     };
 
     const auto layout = gurka::detail::map_to_coordinates(ascii_map, gridsize);
-    map = gurka::buildtiles(layout, ways, {}, {}, "test/data/ignore_access");
+    map = gurka::buildtiles(layout, ways, {}, {}, "test/data/algorithm_selection");
+    map.config.put("mjolnir.traffic_extract", "test/data/algorithm_selection/traffic.tar");
 
     // add live traffic
-    vu::customize_live_traffic_data(map.config, [&](baldr::GraphReader& reader,
-                                                    baldr::TrafficTile& traffic_tile, int edge_index,
-                                                    valhalla::baldr::TrafficSpeed* traffic_speed) {
-      baldr::GraphId edge_id(traffic_tile.header->tile_id);
-      edge_id.set_id(edge_index);
-    });
+    test::build_live_traffic_data(map.config);
+    test::customize_live_traffic_data(map.config,
+                                      [&](baldr::GraphReader& reader,
+                                          baldr::TrafficTile& traffic_tile, int edge_index,
+                                          valhalla::baldr::TrafficSpeed* traffic_speed) {
+                                        baldr::GraphId edge_id(traffic_tile.header->tile_id);
+                                        edge_id.set_id(edge_index);
+                                      });
 
-    // add historical traffic
+    // TODO: add historical traffic
+
+    // std::cout << test::dump_geojson_graph(map.config) << std::endl;
   }
 };
 
@@ -327,7 +329,7 @@ TEST_F(AlgorithmTest, TDReverse) {
 
 // this happens only with non-trivial routes with no date_time or invariant date_time
 TEST_F(AlgorithmTest, Bidir) {
-  {
+  /*{
     auto api = gurka::route(map, {"4", "0"}, "auto");
     ASSERT_EQ(api.trip().routes(0).legs(0).algorithms(0), "bidirectional_a*");
   }
@@ -336,5 +338,5 @@ TEST_F(AlgorithmTest, Bidir) {
     auto api = gurka::route(map, {"A", "2"}, "auto",
                             {{"/date_time/type", "4"}, {"/date_time/value", "2020-10-29T09:00"}});
     ASSERT_EQ(api.trip().routes(0).legs(0).algorithms(0), "bidirectional_a*");
-  }
+  }*/
 }
