+++ test/gurka/test_traffic.cc
@@ -1,10 +1,8 @@
 #include "gurka.h"
-#include "microtar.h"
-#include <gtest/gtest.h>
+#include "test.h"
 
 #include "baldr/graphreader.h"
 #include "baldr/traffictile.h"
-#include "test/util/traffic_utils.h"
 
 #include <boost/property_tree/ptree.hpp>
 
@@ -14,7 +12,7 @@
 #include <sys/stat.h>
 
 using namespace valhalla;
-using LiveTrafficCustomize = valhalla_tests::utils::LiveTrafficCustomize;
+using LiveTrafficCustomize = test::LiveTrafficCustomize;
 
 TEST(Traffic, BasicUpdates) {
 
@@ -35,9 +33,9 @@ TEST(Traffic, BasicUpdates) {
 
   map.config.put("mjolnir.traffic_extract", tile_dir + "/traffic.tar");
 
-  valhalla_tests::utils::build_live_traffic_data(map.config);
+  test::build_live_traffic_data(map.config);
 
-  auto clean_reader = valhalla_tests::utils::make_clean_graphreader(map.config.get_child("mjolnir"));
+  auto clean_reader = test::make_clean_graphreader(map.config.get_child("mjolnir"));
   printf("Do a route with initial traffic");
   {
     auto result = gurka::route(map, "A", "C", "auto", {{"/date_time/type", "0"}}, clean_reader);
@@ -61,7 +59,7 @@ TEST(Traffic, BasicUpdates) {
       current->speed1 = 24 >> 1;
     }
   };
-  valhalla_tests::utils::customize_live_traffic_data(map.config, cb_setter_24kmh);
+  test::customize_live_traffic_data(map.config, cb_setter_24kmh);
 
   printf(" Now do another route with the same (not restarted) actor to see if"
          " it's noticed the changes in the live traffic file");
@@ -84,7 +82,7 @@ TEST(Traffic, BasicUpdates) {
       current->overall_speed = valhalla::baldr::MAX_TRAFFIC_SPEED_KPH >> 1;
     }
   };
-  valhalla_tests::utils::customize_live_traffic_data(map.config, cb_setter_max);
+  test::customize_live_traffic_data(map.config, cb_setter_max);
   {
     auto result = gurka::route(map, "A", "C", "auto", {{"/date_time/type", "0"}}, clean_reader);
     gurka::assert::osrm::expect_steps(result, {"AB"});
@@ -93,7 +91,7 @@ TEST(Traffic, BasicUpdates) {
   }
 
   printf("Back to previous speed");
-  valhalla_tests::utils::customize_live_traffic_data(map.config, cb_setter_24kmh);
+  test::customize_live_traffic_data(map.config, cb_setter_24kmh);
   // And verify that without the "current" timestamp, the live traffic
   // results aren't used
   {
@@ -151,12 +149,11 @@ TEST(Traffic, CutGeoms) {
                                {{"mjolnir.traffic_extract", tile_dir + "/traffic.tar"}});
 
   // empty traffic for now
-  valhalla_tests::utils::build_live_traffic_data(map.config);
+  test::build_live_traffic_data(map.config);
 
   // first we get the edge without traffic on it
   {
-    auto clean_reader =
-        valhalla_tests::utils::make_clean_graphreader(map.config.get_child("mjolnir"));
+    auto clean_reader = test::make_clean_graphreader(map.config.get_child("mjolnir"));
     tyr::actor_t actor(map.config, *clean_reader);
     valhalla::Api api;
     actor.route(
@@ -218,10 +215,9 @@ TEST(Traffic, CutGeoms) {
       }
     };
 
-    valhalla_tests::utils::customize_live_traffic_data(map.config, cb_setter_speed);
+    test::customize_live_traffic_data(map.config, cb_setter_speed);
 
-    auto clean_reader =
-        valhalla_tests::utils::make_clean_graphreader(map.config.get_child("mjolnir"));
+    auto clean_reader = test::make_clean_graphreader(map.config.get_child("mjolnir"));
     tyr::actor_t actor(map.config, *clean_reader);
     valhalla::Api api;
     actor.route(
@@ -294,11 +290,10 @@ TEST(Traffic, CutGeoms) {
         }
       };
 
-      valhalla_tests::utils::customize_live_traffic_data(map.config, cb_setter_speed);
+      test::customize_live_traffic_data(map.config, cb_setter_speed);
     }
 
-    auto clean_reader =
-        valhalla_tests::utils::make_clean_graphreader(map.config.get_child("mjolnir"));
+    auto clean_reader = test::make_clean_graphreader(map.config.get_child("mjolnir"));
     tyr::actor_t actor(map.config, *clean_reader);
     valhalla::Api api;
     actor.route(
@@ -377,11 +372,10 @@ TEST(Traffic, CutGeoms) {
         }
       };
 
-      valhalla_tests::utils::customize_live_traffic_data(map.config, cb_setter_speed);
+      test::customize_live_traffic_data(map.config, cb_setter_speed);
     }
 
-    auto clean_reader =
-        valhalla_tests::utils::make_clean_graphreader(map.config.get_child("mjolnir"));
+    auto clean_reader = test::make_clean_graphreader(map.config.get_child("mjolnir"));
     tyr::actor_t actor(map.config, *clean_reader);
     valhalla::Api api;
     {
@@ -497,11 +491,10 @@ TEST(Traffic, CutGeoms) {
           }
         };
 
-        valhalla_tests::utils::customize_live_traffic_data(map.config, cb_setter_speed);
+        test::customize_live_traffic_data(map.config, cb_setter_speed);
       }
 
-      auto clean_reader =
-          valhalla_tests::utils::make_clean_graphreader(map.config.get_child("mjolnir"));
+      auto clean_reader = test::make_clean_graphreader(map.config.get_child("mjolnir"));
       tyr::actor_t actor(map.config, *clean_reader);
       valhalla::Api api;
       {
@@ -726,10 +719,9 @@ protected:
     closure_map = gurka::buildtiles(layout, ways, {}, {}, tile_dir);
 
     closure_map.config.put("mjolnir.traffic_extract", tile_dir + "/traffic.tar");
-    valhalla_tests::utils::build_live_traffic_data(closure_map.config);
+    test::build_live_traffic_data(closure_map.config);
 
-    clean_reader =
-        valhalla_tests::utils::make_clean_graphreader(closure_map.config.get_child("mjolnir"));
+    clean_reader = test::make_clean_graphreader(closure_map.config.get_child("mjolnir"));
   }
 };
 
@@ -760,7 +752,7 @@ TEST_F(WaypointsOnClosuresTest, DepartPointAtClosure) {
 
       SetLiveSpeed(current, (is_bc ? 0 : default_speed));
     };
-    valhalla_tests::utils::customize_live_traffic_data(closure_map.config, close_edge);
+    test::customize_live_traffic_data(closure_map.config, close_edge);
 
     auto result =
         gurka::route(closure_map, "1", "A", "auto", {{"/date_time/type", "0"}}, clean_reader);
@@ -779,7 +771,7 @@ TEST_F(WaypointsOnClosuresTest, DepartPointAtClosure) {
 
       SetLiveSpeed(current, (is_cb ? 0 : default_speed));
     };
-    valhalla_tests::utils::customize_live_traffic_data(closure_map.config, close_edge);
+    test::customize_live_traffic_data(closure_map.config, close_edge);
 
     auto result =
         gurka::route(closure_map, "1", "A", "auto", {{"/date_time/type", "0"}}, clean_reader);
@@ -803,7 +795,7 @@ TEST_F(WaypointsOnClosuresTest, DepartPointAtClosure) {
 
       SetLiveSpeed(current, ((is_cb || is_bc) ? 0 : default_speed));
     };
-    valhalla_tests::utils::customize_live_traffic_data(closure_map.config, close_edge);
+    test::customize_live_traffic_data(closure_map.config, close_edge);
 
     auto result =
         gurka::route(closure_map, "1", "A", "auto", {{"/date_time/type", "0"}}, clean_reader);
@@ -826,7 +818,7 @@ TEST_F(WaypointsOnClosuresTest, ArrivePointAtClosure) {
 
       SetLiveSpeed(current, (is_da ? 0 : default_speed));
     };
-    valhalla_tests::utils::customize_live_traffic_data(closure_map.config, close_edge);
+    test::customize_live_traffic_data(closure_map.config, close_edge);
 
     auto result =
         gurka::route(closure_map, "B", "2", "auto", {{"/date_time/type", "0"}}, clean_reader);
@@ -845,7 +837,7 @@ TEST_F(WaypointsOnClosuresTest, ArrivePointAtClosure) {
 
       SetLiveSpeed(current, (is_ad ? 0 : default_speed));
     };
-    valhalla_tests::utils::customize_live_traffic_data(closure_map.config, close_edge);
+    test::customize_live_traffic_data(closure_map.config, close_edge);
 
     auto result =
         gurka::route(closure_map, "B", "2", "auto", {{"/date_time/type", "0"}}, clean_reader);
@@ -869,7 +861,7 @@ TEST_F(WaypointsOnClosuresTest, ArrivePointAtClosure) {
 
       SetLiveSpeed(current, ((is_da || is_ad) ? 0 : default_speed));
     };
-    valhalla_tests::utils::customize_live_traffic_data(closure_map.config, close_edge);
+    test::customize_live_traffic_data(closure_map.config, close_edge);
 
     auto result =
         gurka::route(closure_map, "B", "2", "auto", {{"/date_time/type", "0"}}, clean_reader);
@@ -897,7 +889,7 @@ TEST_F(WaypointsOnClosuresTest, IgnoreDepartPointAtClosure) {
                           (CB.Tile_Base() == tile_id && CB.id() == index);
       SetLiveSpeed(current, should_close ? 0 : default_speed);
     };
-    valhalla_tests::utils::customize_live_traffic_data(closure_map.config, close_edge);
+    test::customize_live_traffic_data(closure_map.config, close_edge);
 
     auto result =
         gurka::route(closure_map, "1", "A", "auto", {{"/date_time/type", "0"}}, clean_reader);
@@ -915,7 +907,7 @@ TEST_F(WaypointsOnClosuresTest, IgnoreDepartPointAtClosure) {
                           (CB.Tile_Base() == tile_id && CB.id() == index);
       SetLiveSpeed(current, should_close ? 0 : default_speed);
     };
-    valhalla_tests::utils::customize_live_traffic_data(closure_map.config, close_edge);
+    test::customize_live_traffic_data(closure_map.config, close_edge);
 
     auto result =
         gurka::route(closure_map, "1", "A", "auto",
