+++ test/isochrone.cc
@@ -8,7 +8,6 @@
 #include "thor/worker.h"
 
 #include "test.h"
-#include "utils.h"
 
 using namespace valhalla;
 using namespace valhalla::thor;
