+++ test/mapmatch_config.cc
@@ -1,7 +1,6 @@
 #include "meili/config.h"
 
 #include "test.h"
-#include "utils.h"
 
 namespace {
