+++ test/shape_attributes.cc
@@ -1,5 +1,4 @@
 #include "test.h"
-#include "utils.h"
 
 #include <iostream>
 #include <stdexcept>
