+++ test/test.cc
@@ -1,37 +1,60 @@
-#include "microtar.h"
+#include "test.h"
 
 #include "baldr/graphreader.h"
+#include "baldr/rapidjson_utils.h"
 #include "baldr/traffictile.h"
 #include "mjolnir/graphtilebuilder.h"
 
-#include <boost/property_tree/ptree.hpp>
-
 #include <cmath>
+#include <fstream>
+#include <random>
 #include <sstream>
+#include <stdexcept>
+#include <string>
 #include <sys/mman.h>
 #include <sys/stat.h>
 
-using namespace valhalla;
+#include <gmock/gmock.h>
+#include <gtest/gtest.h>
 
-namespace valhalla_tests {
-namespace utils {
+#include <boost/property_tree/ptree.hpp>
 
-/**
- * Generate a new GraphReader that doesn't re-use a previously
- * statically initizalized tile_extract member variable.
- *
- * Useful if you need to reload a tile extract within the same
- * process
- */
-std::shared_ptr<baldr::GraphReader>
+#include "microtar.h"
+
+namespace test {
+
+std::string load_binary_file(const std::string& filename) {
+  std::string bytes;
+  std::ifstream input_pbf(filename, std::ios::in | std::ios::binary);
+  if (input_pbf.is_open()) {
+    input_pbf.seekg(0, std::ios::end);
+    bytes.resize(input_pbf.tellg());
+    input_pbf.seekg(0, std::ios::beg);
+    input_pbf.read(&bytes[0], bytes.size());
+    input_pbf.close();
+  } else {
+    throw std::runtime_error("Failed to read " + filename);
+  }
+  return bytes;
+}
+
+boost::property_tree::ptree json_to_pt(const std::string& json) {
+  std::stringstream ss;
+  ss << json;
+  boost::property_tree::ptree pt;
+  rapidjson::read_json(ss, pt);
+  return pt;
+}
+
+std::shared_ptr<valhalla::baldr::GraphReader>
 make_clean_graphreader(const boost::property_tree::ptree& mjolnir_conf) {
 
   // Wrapper sub-class to allow replacing the statically initialized
   // tile_extract member variable
-  struct ResettingGraphReader : baldr::GraphReader {
+  struct ResettingGraphReader : valhalla::baldr::GraphReader {
     ResettingGraphReader(boost::property_tree::ptree pt) : GraphReader(pt) {
       // Reset the statically initialized tile_extract_ member variable
-      tile_extract_.reset(new baldr::GraphReader::tile_extract_t(pt));
+      tile_extract_.reset(new valhalla::baldr::GraphReader::tile_extract_t(pt));
     }
   };
   return std::make_shared<ResettingGraphReader>(mjolnir_conf);
@@ -58,7 +81,7 @@ typedef struct {
 //
 /*************************************************************/
 void build_live_traffic_data(const boost::property_tree::ptree& config,
-                             uint32_t traffic_tile_version = valhalla::baldr::TRAFFIC_TILE_VERSION) {
+                             uint32_t traffic_tile_version) {
 
   std::string tile_dir = config.get<std::string>("mjolnir.tile_dir");
   std::string traffic_extract = config.get<std::string>("mjolnir.traffic_extract");
@@ -72,7 +95,7 @@ void build_live_traffic_data(const boost::property_tree::ptree& config,
       throw std::runtime_error("Could not create traffic tar file");
     }
 
-    baldr::GraphReader reader(config.get_child("mjolnir"));
+    valhalla::baldr::GraphReader reader(config.get_child("mjolnir"));
 
     auto tile_ids = reader.GetTileSet();
     // Traffic data works like this:
@@ -87,13 +110,13 @@ void build_live_traffic_data(const boost::property_tree::ptree& config,
     for (auto tile_id : tile_ids) {
       auto tile = reader.GetGraphTile(tile_id);
       std::stringstream buffer;
-      baldr::TrafficTileHeader header = {};
+      valhalla::baldr::TrafficTileHeader header = {};
       header.tile_id = tile_id;
       header.traffic_tile_version = traffic_tile_version;
-      std::vector<baldr::TrafficSpeed> speeds;
+      std::vector<valhalla::baldr::TrafficSpeed> speeds;
       header.directed_edge_count = tile->header()->directededgecount();
       buffer.write(reinterpret_cast<char*>(&header), sizeof(header));
-      baldr::TrafficSpeed dummy_speed = {}; // Initialize to all zeros
+      valhalla::baldr::TrafficSpeed dummy_speed = {}; // Initialize to all zeros
       for (uint32_t i = 0; i < header.directed_edge_count; ++i) {
         buffer.write(reinterpret_cast<char*>(&dummy_speed), sizeof(dummy_speed));
       }
@@ -126,9 +149,6 @@ void build_live_traffic_data(const boost::property_tree::ptree& config,
 // `setter_cb` is a callback that can modify traffic for each edge
 // when building traffic data
 /*************************************************************/
-using LiveTrafficCustomize = std::function<
-    void(baldr::GraphReader&, baldr::TrafficTile&, int, valhalla::baldr::TrafficSpeed*)>;
-
 void customize_live_traffic_data(const boost::property_tree::ptree& config,
                                  const LiveTrafficCustomize& setter_cb) {
   // Now we have the tar-file and can go ahead with per edge customizations
@@ -153,13 +173,13 @@ void customize_live_traffic_data(const boost::property_tree::ptree& config,
 
     // Read every speed tile, and update it with fixed speed of `new_speed` km/h (original speeds are
     // 10km/h)
-    baldr::GraphReader reader(config.get_child("mjolnir"));
+    valhalla::baldr::GraphReader reader(config.get_child("mjolnir"));
     mtar_header_t tar_header;
     while ((mtar_read_header(&tar, &tar_header)) != MTAR_ENULLRECORD) {
-      baldr::TrafficTile tile(reinterpret_cast<char*>(tar.stream) + tar.pos +
-                              sizeof(mtar_raw_header_t_));
+      valhalla::baldr::TrafficTile tile(reinterpret_cast<char*>(tar.stream) + tar.pos +
+                                        sizeof(mtar_raw_header_t_));
 
-      baldr::GraphId tile_id(tile.header->tile_id);
+      valhalla::baldr::GraphId tile_id(tile.header->tile_id);
 
       for (uint32_t index = 0; index < tile.header->directed_edge_count; index++) {
         valhalla::baldr::TrafficSpeed* current =
@@ -173,11 +193,10 @@ void customize_live_traffic_data(const boost::property_tree::ptree& config,
   }
 }
 
-using HistoricalTrafficCustomize = std::function<std::vector<int16_t>(DirectedEdge&)>;
 void customize_historical_traffic(const boost::property_tree::ptree& config,
                                   const HistoricalTrafficCustomize& cb) {
   // loop over all tiles in the tileset
-  baldr::GraphReader reader(config.get_child("mjolnir"));
+  valhalla::baldr::GraphReader reader(config.get_child("mjolnir"));
   auto tile_dir = config.get<std::string>("mjolnir.tile_dir");
   for (const auto& tile_id : reader.GetTileSet()) {
     valhalla::mjolnir::GraphTileBuilder tile(tile_dir, tile_id, false);
@@ -193,5 +212,62 @@ void customize_historical_traffic(const boost::property_tree::ptree& config,
   }
 }
 
-} // namespace utils
-} // namespace valhalla_tests
+// TODO: bidirectional edges overlap shapes so its hard to visualize them
+// just draw one shape for both edges and add both edges properties
+std::string dump_geojson_graph(const boost::property_tree::ptree& config) {
+  // flesh out the geojson to start
+  rapidjson::Document doc(rapidjson::kObjectType);
+  doc.AddMember("type", "FeatureCollection", doc.GetAllocator());
+  rapidjson::Value features(rapidjson::kArrayType);
+
+  // for all tiles for all edges
+  valhalla::baldr::GraphReader reader(config.get_child("mjolnir"));
+  for (auto tile_id : reader.GetTileSet()) {
+    if (reader.OverCommitted())
+      reader.Trim();
+    const auto* tile = reader.GetGraphTile(tile_id);
+    for (const auto& edge : tile->GetDirectedEdges()) {
+      valhalla::baldr::GraphId edge_id(tile_id.tileid(), tile_id.level(),
+                                       &edge - tile->directededge(0));
+      auto info = tile->edgeinfo(edge.edgeinfo_offset());
+
+      // add some properties
+      rapidjson::Value properties(rapidjson::kObjectType);
+      rapidjson::Value names(rapidjson::kArrayType);
+      for (const std::string& name : info.GetNames()) {
+        names.PushBack(rapidjson::Value(name, doc.GetAllocator()).Move(), doc.GetAllocator());
+      }
+      properties.AddMember("edge_id", std::to_string(edge_id), doc.GetAllocator());
+      properties.AddMember("names", names, doc.GetAllocator());
+
+      // add the geom
+      rapidjson::Value geometry(rapidjson::kObjectType);
+      geometry.AddMember("type", "LineString", doc.GetAllocator());
+      rapidjson::Value coordinates(rapidjson::kArrayType);
+      for (const auto& point : info.shape()) {
+        rapidjson::Value coordinate(rapidjson::kArrayType);
+        coordinate.PushBack(point.lng(), doc.GetAllocator());
+        coordinate.PushBack(point.lat(), doc.GetAllocator());
+        coordinates.PushBack(coordinate, doc.GetAllocator());
+      }
+      geometry.AddMember("coordinates", coordinates, doc.GetAllocator());
+
+      // keep all the stuff we made
+      rapidjson::Value feature(rapidjson::kObjectType);
+      feature.AddMember("type", "Feature", doc.GetAllocator());
+      feature.AddMember("properties", properties, doc.GetAllocator());
+      feature.AddMember("geometry", geometry, doc.GetAllocator());
+      features.PushBack(feature, doc.GetAllocator());
+    }
+  }
+
+  // serialize
+  doc.AddMember("features", features, doc.GetAllocator());
+  rapidjson::StringBuffer sb;
+  rapidjson::PrettyWriter<rapidjson::StringBuffer> writer(sb);
+  doc.Accept(writer);
+  return sb.GetString();
+}
+
+} // namespace test
+
