+++ test/test.h
@@ -1,15 +1,24 @@
 // -*- mode: c++ -*-
+#pragma once
 
-#ifndef TEST_HPP
-#define TEST_HPP
+#include "baldr/graphreader.h"
+#include "baldr/rapidjson_utils.h"
+#include "baldr/traffictile.h"
+#include "mjolnir/graphtilebuilder.h"
 
+#include <cmath>
 #include <fstream>
 #include <random>
+#include <sstream>
 #include <stdexcept>
 #include <string>
+#include <sys/mman.h>
+#include <sys/stat.h>
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
+
+#include <boost/property_tree/ptree.hpp>
 
 namespace test {
 
@@ -19,20 +28,7 @@ inline float rand01(std::mt19937& gen) {
   return dis(gen);
 }
 
-inline std::string load_binary_file(const std::string& filename) {
-  std::string bytes;
-  std::ifstream input_pbf(filename, std::ios::in | std::ios::binary);
-  if (input_pbf.is_open()) {
-    input_pbf.seekg(0, std::ios::end);
-    bytes.resize(input_pbf.tellg());
-    input_pbf.seekg(0, std::ios::beg);
-    input_pbf.read(&bytes[0], bytes.size());
-    input_pbf.close();
-  } else {
-    throw std::runtime_error("Failed to read " + filename);
-  }
-  return bytes;
-}
+std::string load_binary_file(const std::string& filename);
 
 MATCHER_P2(IsBetween,
            a,
@@ -46,6 +42,47 @@ template <typename pbf_message_t> bool pbf_equals(const pbf_message_t& a, const 
   return a.SerializeAsString() == b.SerializeAsString();
 }
 
-} // namespace test
+boost::property_tree::ptree json_to_pt(const std::string& json);
 
-#endif
+/**
+ * Generate a new GraphReader that doesn't re-use a previously
+ * statically initizalized tile_extract member variable.
+ *
+ * Useful if you need to reload a tile extract within the same
+ * process
+ */
+std::shared_ptr<valhalla::baldr::GraphReader>
+make_clean_graphreader(const boost::property_tree::ptree& mjolnir_conf);
+
+/*************************************************************/
+// Creates an empty traffic file
+//
+// To actually customize the traffic data, use `customize_live_traffic_data`
+//
+/*************************************************************/
+void build_live_traffic_data(const boost::property_tree::ptree& config,
+                             uint32_t traffic_tile_version = valhalla::baldr::TRAFFIC_TILE_VERSION);
+
+/*************************************************************/
+// Helper function for customizing traffic data in unit-tests
+//
+// `setter_cb` is a callback that can modify traffic for each edge
+// when building traffic data
+/*************************************************************/
+using LiveTrafficCustomize = std::function<void(valhalla::baldr::GraphReader&,
+                                                valhalla::baldr::TrafficTile&,
+                                                int,
+                                                valhalla::baldr::TrafficSpeed*)>;
+
+void customize_live_traffic_data(const boost::property_tree::ptree& config,
+                                 const LiveTrafficCustomize& setter_cb);
+
+using HistoricalTrafficCustomize = std::function<std::vector<int16_t>(DirectedEdge&)>;
+void customize_historical_traffic(const boost::property_tree::ptree& config,
+                                  const HistoricalTrafficCustomize& cb);
+
+// TODO: bidirectional edges overlap shapes so its hard to visualize them
+// just draw one shape for both edges and add both edges properties
+std::string dump_geojson_graph(const boost::property_tree::ptree& config);
+
+} // namespace test
